// src/module/entities/model/Client.ts
export class Client {
  sigeemp: string;
  clienteCod: number;
  credito: number;
  razaoSocial: string;
  nomeFantasia: string;
  tipo: string;
  cnpj: number;
  ie: string;
  im: string;
  celular: number;
  telefone: number;
  telefone2: number;
  fax: number;
  ativo: number;
  conceitoCod: number;
  grupoCadCod: number;
  suframa: number;
  transportadoraCod: number;
  holding: number;
  uf: string;
  cidadeIbgeCod: number;
  cidade: string;
  bairro: string;
  logradouro: string;
  numero: number;
  complemento: string;
  cep: number;
  locCobId: number;
  dda: number;
  dtCadastro: Date;
  dtAlteracao: Date;
  hrAlteracao: number;
}
export class ClientFields {
  sigeemp?: boolean;
  clienteCod?: boolean;
  credito?: boolean;
  razaoSocial?: boolean;
  nomeFantasia?: boolean;
  tipo?: boolean;
  cnpj?: boolean;
  ie?: boolean;
  im?: boolean;
  celular?: boolean;
  telefone?: boolean;
  telefone2?: boolean;
  fax?: boolean;
  ativo?: boolean;
  conceitoCod?: boolean;
  grupoCadCod?: boolean;
  suframa?: boolean;
  transportadoraCod?: boolean;
  holding?: boolean;
  uf?: boolean;
  cidadeIbgeCod?: boolean;
  cidade?: boolean;
  bairro?: boolean;
  logradouro?: boolean;
  numero?: boolean;
  complemento?: boolean;
  cep?: boolean;
  locCobId?: boolean;
  dda?: boolean;
  dtCadastro?: boolean;
  dtAlteracao?: boolean;
  hrAlteracao?: boolean;
}